Add files via upload
Welcome to the Super Store Sales Analysis GitHub repository! This project focuses on utilizing Power BI visualizations to analyse the sales data of Superstore, a store with branches in USA. The repository contains a collection of visualizations that provide insights into sales patterns at various geographic levels and delve into various category of products.

OBJECTIVE:
The primary objectives of this project are as follows:
1.	Perform Basic Data Cleaning:
	Ensure data quality and accuracy by performing necessary cleaning operations on the dataset.

2.	Segment Visuals Using State, Region, category and city.
	Provide segmented visualizations to allow stakeholders to analyse sales data.

3.	Generate Reports for Sales Over Cities, States, Regions, and Markets:
	Create comprehensive reports that showcase sales performance over different geographic levels, including cities, states, regions, and markets.

4.	Create Interactive Sales Report for Stakeholders:
	Develop an interactive Power BI report that allows stakeholders to explore and understand sales data dynamically.

VISUALISATIONS:
1. Sales by category : Pie Chart
2. Sales by sub-category : Funnel Chart
3. Sales by region : Donut Chart
4. Sales by state : Map
5. Sales by date : Line Chart
6. Sum of quantity, discount, profit, sales : Cards

DATA SOURCE:
This project uses the Super Store dataset, which is open source and available on Kaggle. You can access the dataset here. The dataset is in CSV format with 9995 rows and 21 columns.

CONCLUSION:
The analysis and visualizations provided in this project aim to offer stakeholders a comprehensive understanding of Super Store's sales data. By exploring different aspects of the dataset, users can gain valuable insights into sales patterns, shipping costs, and geographic performance.